# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Discord Clone API
  version: 1.0.0
  description: |
    API REST pour le clone Discord.
    
    ## Authentification
    Toutes les routes (sauf `/auth/register` et `/auth/login`) nécessitent un token JWT.
    
    ```
    Authorization: Bearer <access_token>
    ```
  contact:
    name: Discord Clone Team
    email: team@discord-clone.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development
  - url: https://api.discord-clone.example.com/api/v1
    description: Production
...
paths:
...
  /servers/{serverId}/members:
    parameters:
      - $ref: '#/components/parameters/ServerId'
    get:
      tags: [Members]
      summary: List server members
      operationId: listMembers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /servers/{serverId}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/ServerId'
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [Members]
      summary: Get member details
      operationId: getMember
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Member details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Members]
      summary: Update member role
      operationId: updateMemberRole
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Member role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Members]
      summary: Kick member
      operationId: kickMember
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Member kicked
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ServerId:
      name: serverId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    Role:
      type: string
      enum: [Owner, Admin, Moderator, Member]
      description: "Member role in PascalCase (matches Rust enum serialization)"
...
    Member:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        server_id:
          type: string
          format: uuid
        username:
          type: string
        avatar_url:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/Role'
        joined_at:
          type: string
          format: date-time
    MemberDetails:
      allOf:
        - $ref: '#/components/schemas/Member'
        - type: object
          properties:
            email:
              type: string
              format: email
            permissions:
              type: array
              items:
                type: string
    MemberList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Member'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    UpdateRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [Admin, Moderator, Member]
          description: "Cannot assign Owner role via this endpoint"
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer